requirements.txt
numpy>=1.26
pandas>=2.1
pyarrow>=14.0
PyYAML>=6.0
tqdm>=4.66
optuna>=3.6
torch>=2.2
